pyproject.toml
[tool.poetry]
name = "covid_19"
version = "0.1.0"
description = "Simple module to fetch data about COVI-19 cases in Brasil and store in a SQLite database."
authors = ["Weverton Marques <wevertonms@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
pandas = "^1.0.3"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
